M3 Custom Data Class Assignment
A class is a datatype so we would categorize it with things like int, boolean, string, and so on. It's just a better data type designed to fit our needs explicitly. Choose a real world object, the items on the list below are suggestions based off of previous student's personal interests, and create a class to represent object as a single digital datatype: be sure to include constructor(s), make your datafields mangled and include all getters and setters, and override the __str__() and __eq__() method to get full points. Then create a main class/script to instantiate a couple unique instances of the class and print them to the console (the main must be in a separate file from the data class).

(I tried to match several previous students' personal interests I know about here, if you have something in mind that would be good for this assignment check with me about creating a data class to represent it)
NOTE: The real world object you are making a virtual representation of to be used in software must be approved in advance, send a message early in the module letting me know what you would like to make.
HINT: To come up with something to make it could be helpful to think of either an online retail website or a VR simulation, what qualities and actions are necessary to accurately represent the object in a piece of software?