M6 CI/CD Pipeline, Docker
The purpose of this assignment is to get some practice with modern Software Development tools by creating a CI/CD Pipeline.

A Continuous-Integration/Continuous-Deployment Pipeline is an automated process which is used as a replacement for the traditional Version-Release-Patch numbering system for periodic updates to software. Rather than waiting for the next update the end user simply receives the latest version of the software from the repository every time they run the application.